chore: move scoreboard in his feature folder & clean
