+++ src/components/features/scoreboard/scoreboard.tsx
